var packing

+++ contracts/Generator/NFTTicketGenerator.sol
@@ -25,7 +25,7 @@ contract NFTTicket is NoobFriendlyTokenTemplate {
 
     function initialize(
         string calldata baseURI_,
-        uint160 startTimestamp_,
+        uint128 startTimestamp_,
         uint48[] calldata ticketAmounts_,
         uint160[] calldata ticketPrices_
     ) external onlyOwner onlyOnce {
